tests + added id to Graph interface
