--- src/backend/src/routes/graph.routes.ts
@@ -1,8 +1,0 @@
-import GraphController from '../controllers/graph.controllers'
-import express from 'express';
-
-const graphsRouter = express.Router();
-
-graphsRouter.get('/statistics/:graphId', GraphController.getSingleGraph);
-
-export default graphsRouter;
